requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
Automat==20.2.0
black==20.8b1
boto3==1.17.1
botocore==1.20.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cssselect==1.1.0
Django==3.1.5
django-environ==0.4.5
django-fsm==2.7.1
django-htmx==1.0.1
django-money==1.3
django-rest-framework==0.1.0
django-storages==1.11.1
djangorestframework==3.12.2
drf-yasg==1.20.0
gunicorn==20.0.4
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflection==0.5.1
itemadapter==0.2.0
itemloaders==1.0.4
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
lxml==4.6.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
packaging==20.9
parsel==1.6.0
pathspec==0.8.1
Pillow==8.1.0
priority==1.3.0
Protego==0.1.16
psycopg2-binary==2.8.6
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
queuelib==1.5.0
regex==2020.11.13
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
s3transfer==0.3.4
Scrapy==2.5.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
w3lib==1.22.0
zope.interface==5.3.0
